70: Increase the number of pins from 255 to the next byte.
Task-Url: https://github.com/meduzapat/LEDSpicer/issues/70

* Changed the number of elements and leds from 1 byte to 2 bytes.
* Improved documentation.
* Fixed an issue with the sockets not sending the terminator.
* Fixed some issues with actors.
* Fixed a posible bug while changing profiles.
* Code clen up and documentation fixes.

+++ src/animations/Actor.hpp
@@ -168,9 +168,9 @@ protected:
 	static uint8_t frame;
 
 	uint16_t
-		/// Number of secods to wait until start processing this actor.
+		/// Number of seconds to wait until start processing this actor.
 		secondsToStart = 0,
-		/// Number of secods to wait to stop this actor.
+		/// Number of seconds to wait to stop this actor.
 		secondsToEnd   = 0;
 
 	/// Start time clock for this actor.
@@ -189,7 +189,7 @@ protected:
 	/**
 	 * @return the number of elements on the animation's group.
 	 */
-	uint8_t getNumberOfElements() const;
+	uint16_t getNumberOfElements() const;
 
 	/**
 	 * Changes an element color with a new one applying filters.
@@ -199,7 +199,7 @@ protected:
 	 * @param filter
 	 * @param percent the amount of effect to apply, only used by Combine.
 	 */
-	void changeElementColor(uint8_t index, const Color& color, Color::Filters filter, uint8_t percent = 50);
+	void changeElementColor(uint16_t index, const Color& color, Color::Filters filter, uint8_t percent = 50);
 
 	/**
 	 * Changes the colors with a new one applying filters.
@@ -220,7 +220,7 @@ protected:
 	/**
 	 * @return true if an element got affected (dirty/changed) on this frame.
 	 */
-	bool isElementAffected(uint8_t index);
+	bool isElementAffected(uint16_t index);
 
 	/**
 	 * Checks if the actor will repeat.
